Implemnted starter versions of core execution fns
I need to fix how address calculation is done, its the most
important function in `Core` but I haven't implemented it. classic. The
new program counters are returned in a `CoreEvent` which I'm a bit of a
fan of. I'm wondering if make all of the exec_* functions immutable. It
might not be worth the performance loss though