Address encasulates modulo logic -- stubs
This commit will break all of the tests, `Address` has been rewritten as
a struct that will take a maximum size and wrap around that size. I've
written stubs for the `Add`, `Sub` and other traits and the `*Assign`
versions of those traits for `Field`. Hopefully this will make address
calculate code a lot cleaner